+++ CHANGES.markdown
@@ -1,5 +1,6 @@
 ## next
   - Omit empty `describe` groups from report
+  - Do not print empty lines before/after spec groups
 
 ## Changes in 1.11.4
   - Make test suite independent from QuickCheck seed (see #187)